refactor(query): extract mem stat
