+++ src/common/base/src/runtime/memory/mod.rs
@@ -1,0 +1,5 @@
+mod mem_stat;
+mod stat_buffer;
+
+pub use mem_stat::MemStat;
+pub use stat_buffer::StatBuffer;
